Updated sample according to comments
