--- functions/https/actions.js
@@ -1,14 +1,0 @@
-const express = require('express');
-const baseExpress = require('./base-express.js');
-
-module.exports = {
-  helloWorld: function () {
-    const helloWorldExpress = baseExpress.createBaseExpress();
-
-    helloWorldExpress.get('/', (req, res) => {
-      res.status(200).send("Hello World")
-    });
-
-    return helloWorldExpress;
-  }
-}
